+++ packages/types/src/components/col/col.ts
@@ -1,0 +1,9 @@
+import type { BoxProps } from "../box/box";
+import type { NubeComponent } from "../types";
+
+export type ColProps = Omit<BoxProps, "direction">;
+
+export type Col = NubeComponent &
+	ColProps & {
+		type: "col";
+	};
