+++ packages/types/src/components/field/field.ts
@@ -1,14 +1,17 @@
 import type {
+	NubeComponent,
+	NubeComponentProps,
 	NubeSDKState,
+	UIValue,
 } from "@tiendanube/nube-sdk-types";
 
 export type FieldEventHandler = (data: {
 	type: string;
 	state: NubeSDKState;
-	value?: string;
+	value?: UIValue;
 }) => void;
 
-export type FieldProps = {
+export type FieldProps = NubeComponentProps & {
 	name: string;
 	label: string;
 	onchange?: FieldEventHandler;
@@ -16,6 +19,7 @@ export type FieldProps = {
 	onfocus?: FieldEventHandler;
 };
 
-export type Field = FieldProps & {
-	type: "field";
-};
+export type Field = NubeComponent &
+	FieldProps & {
+		type: "field";
+	};
