--- packages/types/src/ui/col/col.ts
@@ -1,7 +1,0 @@
-import type { BoxProps } from "../box/box";
-
-export type ColProps = Omit<BoxProps, "direction">;
-
-export type Col = ColProps & {
-	type: "col";
-};
