--- packages/types/src/ui/row/row.ts
@@ -1,7 +1,0 @@
-import type { BoxProps } from "../box/box";
-
-export type RowProps = Omit<BoxProps, "direction">;
-
-export type Row = RowProps & {
-	type: "row";
-}
