+++ packages/ui/src/components/box/box.ts
@@ -1,4 +1,4 @@
-import { Box, BoxProps } from "@tiendanube/nube-sdk-types";
+import type { Box, BoxProps } from "@tiendanube/nube-sdk-types";
 
 export const box = (props: BoxProps): Box => ({
 	type: "box",
