+++ packages/ui/src/components/col/col.ts
@@ -1,4 +1,4 @@
-import { Col, ColProps } from "@tiendanube/nube-sdk-types";
+import type { Col, ColProps } from "@tiendanube/nube-sdk-types";
 
 export const col = (props: ColProps): Col => ({
 	type: "col",
