+++ packages/ui/src/components/field/field.ts
@@ -1,6 +1,6 @@
-import { Field, FieldProps } from "@tiendanube/nube-sdk-types";
+import type { Field, FieldProps } from "@tiendanube/nube-sdk-types";
 
 export const field = (props: FieldProps): Field => ({
 	type: "field",
-	...props
+	...props,
 });
