+++ packages/ui/src/components/row/row.ts
@@ -1,4 +1,4 @@
-import { Row, RowProps } from "@tiendanube/nube-sdk-types";
+import type { Row, RowProps } from "@tiendanube/nube-sdk-types";
 
 export const row = (props: RowProps): Row => ({
 	type: "row",
